Encoder Debounce circuit
While I was building and testing this with a variety of flying wires and proto boards, the rotary encoder worked perfectly fine. After building the real circuit and soldering it up with short wires, it didn't have all the stray capacitance which served to debounce the switches... And I had no control over the fan speed.

I searched the Internet and found several examples of rotary encoder debounce circuits that were all the same, so I took that as recommendations, and implemented it on my board, per these woefully un-attributed, stolen from the hardworking people who took the time to draw them up. Sorry! Not sorry! 🤔😂